+++ light-desk/src/frontend/components/rect.tsx
@@ -5,15 +5,38 @@ import * as proto from '../../shared/proto';
 
 function colorToCss(color: proto.ColorJSON) {
   switch (color.type) {
-    case 'rgb':
-      return `rgb(${color.r}, ${color.g}, ${color.b})`;
+    case 'rgba':
+      return `rgba(${color.r}, ${color.g}, ${color.b}, ${color.a})`;
   }
 }
+
+const TRANSPARENCY_SVG = `
+<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 10 10">
+  <rect width="10px" height="10px" fill="#333" />
+  <rect width="5px" height="5px" fill="#ddd" y="5"/>
+  <rect width="5px" height="5px" fill="#ddd" x="5"/>
+</svg>
+`;
 
 interface Props {
   className?: string;
   info: proto.RectComponent;
 }
+
+const Wrapper = styled.div`
+  width: 30px;
+  height: 30px;
+  border-radius: 3px;
+  overflow: hidden;
+  background: url('data:image/svg+xml,${encodeURIComponent(TRANSPARENCY_SVG)}');
+  background-repeat: repeat;
+  background-size: 10px;
+`;
+
+const Inner = styled.div`
+  width: 100%;
+  height: 100%;
+`;
 
 class Rect extends React.Component<Props, never> {
   public constructor(props: Props) {
@@ -22,18 +45,11 @@ class Rect extends React.Component<Props, never> {
 
   public render() {
     return (
-      <div
-        className={this.props.className}
-        style={{ backgroundColor: colorToCss(this.props.info.color) }}
-      />
+      <Wrapper className={this.props.className}>
+        <Inner style={{ backgroundColor: colorToCss(this.props.info.color) }} />
+      </Wrapper>
     );
   }
 }
 
-const StyledRect = styled(Rect)`
-  width: 30px;
-  height: 30px;
-  border-radius: 3px;
-`;
-
-export { StyledRect as Rect };
+export { Rect };