+++ light-desk/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Throttle how often the tree is updated
 - Introduce new color constant `COLOR_RGB_BLACK`
 - Fixed bugs that would occur when a child added to a group multiple times
+- [breaking]: start using Color utilities from `@synesthesia-project/compositor`
 
 ## [2.1.1] - 2022-01-11